# spring-openapi/src/main/resources/springopenapi.yml
openapi: 3.0.0
info:
  title: Spring OpenAPI
  version: 1.0.0
  contact: {}
security: []
servers: []
paths:
  /account/auth:
    post:
      summary: Authentication
      operationId: authenticate
      tags:
        - authentication
      description: ''
      parameters: []
      responses:
        '200':
          description: Authentication Successful
          content:
            application/json:
              schema:
                type: string
        '403':
          description: Access Denied
          content:
            application/json:
              schema: 
                type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthUser'
  '/employee/{id}':
    get:
      summary: Employee by ID
      operationId: getEmployeeById
      tags:
        - employees
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
  /employee:
    get:
      summary: Employee List
      operationId: getEmployeeList
      tags:
        - employees
      description: ''
      parameters: []
      responses:
        '200':
          description: Employee List
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeList'
components:
  links: {}
  callbacks: {}
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: string
        lname:
          type: string
        email:
          type: string
        age:
          type: integer
        experience:
          type: number
        salary:
          type: number
        fname:
          type: string
        gender:
          type: string
        phone:
          type: string
        job_title:
          type: string
        department:
          type: string
    EmployeeList:
      type: array
      items:
        $ref: '#/components/schemas/Employee'
    AuthUser:
      type: object
      properties:
        username:
          type: string
        passwd:
          type: string